% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.sgd.R
\name{coef.sgd}
\alias{coef.sgd}
\title{Extract Model Coefficients}
\usage{
\method{coef}{sgd}(object, ...)
}
\arguments{
\item{object}{object of class \code{sgd}.}

\item{\dots}{some methods for this generic require additional
arguments. None are used in this method.}
}
\value{
Coefficients extracted from the model object \code{object}.
}
\description{
Extract model coefficients from \code{sgd} objects. \code{coefficients}
is an \emph{alias} for it.
}
