% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sGD.R
\name{distmat}
\alias{distmat}
\title{Calculate a pairwise landscape distance matrix (Euclidian or cost-distance).}
\usage{
distmat(sp_points, method = c("ed", "cd"), file_name = NULL,
  landscape = NULL)
}
\arguments{
\item{sp_points}{An object of class SpatialPoints (see raster package).}

\item{method}{Specify the type of distance matrix to be produced, using "ed" for Euclidean distance and "cd" for cost-weighted (i.e. effective) distance or c("ed","cd") for both. If you calculate cost-weighted distances, make sure the \code{points} projection is the same as the \code{landscape} raster.}

\item{file_name}{A character string that will be appended to the beginning of the output filename.}

\item{landscape}{An object of class RasterLayer (see raster package)}
}
\value{
An NxN (N= sample size: i.e. nrow(xy)) matrix of pairwise Euclidean and/or effective landscape distances written to .csv comma delimited files with edmat or cdmat appended to the end of the \code{filename}.
}
\description{
Calculate a pairwise landscape distance matrix (Euclidian or cost-distance).
}
\examples{
library(sGD)
setwd("C:/Temp")
file_name <- "sGD_demo"
xy_file <- system.file("extdata","sGD_demo_xy.csv",package="sGD")
proj <- "+proj=utm +zone=10 +datum=NAD83"
sp_points <- SpatialPoints(read.csv(xy_file)[,c(2,3)],proj4string=CRS(proj))
landscape_ascii <- system.file("extdata","sGD_demo_IBR_landscape.asc",package="sGD")
landscape <- raster(landscape_ascii,crs=CRS(proj))
distmat(sp_points,file_name,"ed")
distmat(sp_points,file_name,"cd",landscape)
distmat(sp_points,file_name,c("ed","cd"),landscape)
}

