% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.sgd.R
\name{fitted.sgd}
\alias{fitted.sgd}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{sgd}(object, ...)
}
\arguments{
\item{object}{object of class \code{sgd}.}

\item{\dots}{some methods for this generic require additional
arguments. None are used in this method.}
}
\value{
Fitted values extracted from the object \code{object}.
}
\description{
Extract fitted values from from \code{sgd} objects.
\code{fitted.values} is an \emph{alias} for it.
}
