\name{LinearRegression}
\alias{LinearRegression}
\title{Linear Regression Constructor}
\usage{
  LinearRegression(...)
}
\arguments{
  \item{...}{named parameters to pass to the constructor}
}
\value{
  LinearRegression object
}
\description{
  Returns an instance of LinearRegression with supplied
  parameters or defaults
}
\author{
  Jake Burkhead <jlburkhead@ucdavis.edu>
}

