\name{LogisticRegression}
\alias{LogisticRegression}
\title{Logistic Regression Constructor}
\usage{
  LogisticRegression(...)
}
\arguments{
  \item{...}{named parameters to pass to the constructor}
}
\value{
  LogisticRegression object
}
\description{
  Returns an instance of LogisticRegression with supplied
  parameters or defaults.
}
\author{
  Jake Burkhead <jlburkhead@ucdavis.edu>
}

