% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sgd.R
\name{plot.sgd}
\alias{plot.sgd}
\title{Generic function for plotting of \code{sgd} objects.}
\usage{
\method{plot}{sgd}(x, type = "mse", ...)
}
\arguments{
\item{x}{sgd object}

\item{type}{character specifying the type of plot: \code{"mse"}}

\item{\dots}{}
}
\description{
Generic function for plotting of \code{sgd} objects.
}

