\name{PoissonRegression}
\alias{PoissonRegression}
\title{Poisson Regression Constructor}
\usage{
  PoissonRegression(...)
}
\arguments{
  \item{...}{named arguments to pass to the constructor}
}
\value{
  PoissonRegression object
}
\description{
  Returns an instance of PoissonRegression with supplied
  parameters or defaults
}
\author{
  Jake Burkhead <jlburkhead@ucdavis.edu>
}

