% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sgd.R
\name{print.sgd}
\alias{print.sgd}
\title{Generic function for printing of \code{sgd} objects.}
\usage{
\method{print}{sgd}(x)
}
\arguments{
\item{x}{sgd object}
}
\description{
Generic function for printing of \code{sgd} objects.
}

