% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.sgd.R
\name{residuals.sgd}
\alias{residuals.sgd}
\title{Extract Model Residuals}
\usage{
\method{residuals}{sgd}(object, ...)
}
\arguments{
\item{object}{object of class \code{sgd}.}

\item{\dots}{some methods for this generic require additional
arguments. None are used in this method.}
}
\value{
Residuals extracted from the object \code{object}.
}
\description{
Extract model residuals from \code{sgd} objects. \code{resid} is an
\emph{alias} for it.
}
