% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sgd.R
\name{plot.sgd}
\alias{plot.sgd}
\title{Plot objects of class \code{sgd}.}
\usage{
\method{plot}{sgd}(x, ..., type = "mse")
}
\arguments{
\item{x}{object of class \code{sgd}.}

\item{type}{character specifying the type of plot: \code{"mse"},
  \code{"runtime"}}

\item{\dots}{}
}
\description{
Plot objects of class \code{sgd}.
}

