% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sgd.R
\name{plot.sgd}
\alias{plot.sgd}
\alias{plot.list}
\title{Plot objects of class \code{sgd}.}
\usage{
\method{plot}{sgd}(x, ..., type = "mse", xaxis = "iteration")

\method{plot}{list}(x, ..., type = "mse", xaxis = "iteration")
}
\arguments{
\item{x}{object of class \code{sgd}.}

\item{\dots}{additional arguments used for each type of plot. See
\sQuote{Details}.}

\item{type}{character specifying the type of plot: \code{"mse"},
\code{"clf"}, \code{"mse-param"}. See \sQuote{Details}. Default is
\code{"mse"}.}

\item{xaxis}{character specifying the x-axis of plot: \code{"iteration"}
plots the y values over the log-iteration of the algorithm;
\code{"runtime"} plots the y values over the time in seconds to reach them.
Default is \code{"iteration"}.}
}
\description{
Plot objects of class \code{sgd}.
}
\details{
Types of plots available:
\describe{
  \item{\code{mse}}{Mean squared error in predictions, which takes the
    following arguments:
    \describe{
      \item{\code{x_test}}{test set}
      \item{\code{y_test}}{test responses to compare predictions to}
    }}
  \item{\code{clf}}{Classification error in predictions, which takes the
    following arguments:
    \describe{
      \item{\code{x_test}}{test set}
      \item{\code{y_test}}{test responses to compare predictions to}
    }}
  \item{\code{mse-param}}{Mean squared error in parameters, which takes the
    following arguments:
    \describe{
      \item{\code{true_param}}{true vector of parameters to compare to}
    }}
}
}
