% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.sgd.R
\name{plot.sgd}
\alias{plot.list}
\alias{plot.sgd}
\title{Plot objects of class \code{sgd}.}
\usage{
\method{plot}{sgd}(x, ..., type = "mse")

\method{plot}{list}(x, ..., type = "mse")
}
\arguments{
\item{x}{object of class \code{sgd}.}

\item{type}{character specifying the type of plot: \code{"mse"},
  \code{"runtime"}. See \sQuote{Details}.}

\item{\dots}{additional arguments used for each type of plot. See
\sQuote{Details}.}
}
\description{
Plot objects of class \code{sgd}.
}
\details{
Types:
\describe{
  \item{\code{mse}}{Mean squared error in predictions, which takes the
    following arguments:
    \describe{
      \item{\code{x_test}}{test set}
      \item{\code{y_test}}{test responses to compare predictions to}
    }}
  \item{\code{runtime}}{Mean squared error in predictions over runtime, which
    takes the following arguments:
    \describe{
      \item{\code{x_test}}{test set}
      \item{\code{y_test}}{test responses to compare predictions to}
    }}
  \item{\code{mse-param}}{Mean squared error in parameters, which takes the
    following arguments:
    \describe{
      \item{\code{true_param}}{true vector of parameters to compare to}
    }}
  \item{\code{mse-param-runtime}}{Mean squared error in parameters over
    runtime, which takes the following arguments:
    \describe{
      \item{\code{true_param}}{true vector of parameters to compare to}
    }}
}
}

