% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.sgd.R
\name{predict.sgd}
\alias{predict.sgd}
\alias{predict_all}
\title{Predict for objects of class \code{sgd}}
\usage{
\method{predict}{sgd}(object, x_test, ...)

predict_all(object, x_test, ...)
}
\arguments{
\item{object}{object of class \code{sgd}.}

\item{x_test}{design matrix to form predictions on}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Form predictions using the estimated model parameters from stochastic
gradient descent.
}
\details{
A column of 1's must be included if the parameters include a bias, or
intercept, term.
}

