% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sgd.R
\name{predict.sgd}
\alias{predict.sgd}
\title{Form predictions using the estimated model parameters from stochastic
gradient descent.}
\usage{
\method{predict}{sgd}(x, x_test, ...)
}
\arguments{
\item{x}{object of class \code{sgd}.}

\item{x_test}{design matrix to form predictions on}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Form predictions using the estimated model parameters from stochastic
gradient descent.
}

