% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.sgd.R
\name{print.sgd}
\alias{print.sgd}
\title{Print objects of class \code{sgd}.}
\usage{
\method{print}{sgd}(x, ...)
}
\arguments{
\item{x}{object of class \code{sgd}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Print objects of class \code{sgd}.
}

