\name{sgd-package}
\alias{sgd-package}
\alias{sgd}
\docType{package}
\title{
  (Stochastic)? Gradient Descent
}
\description{
  Toy package to mess around with using Rcpp and RcppArmadillo to link c++ and R. Implements stochastic and batch gradient descent in c++.
}
\details{
\tabular{ll}{
Package: \tab sgd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-28\cr
License: \tab BSD\cr
}
}
\author{
Jake Burkhead

Maintainer: Jake Burkhead <jlburkhead@ucdavis.edu>
}
\references{
  Machine Learning by Kevin Murphy
  http://www.cs.toronto.edu/~hinton/csc2515/notes/lec6tutorial.pdf
}
\keyword{ package }
\seealso{
}
\examples{
}
