% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sGD.R
\name{sGD}
\alias{sGD}
\title{Calculate spatially explicit indicies of genetic diversity and Wright's neighborhood size (NS).}
\usage{
sGD(genepop_file, output_name, xy_file, dist_mat, radius, min_N, NS_ans, GD_ans,
  NeEstimator_dir = NULL)
}
\arguments{
\item{genepop_file}{The path to a genotype file in genepop format (with a .gen extension).}

\item{output_name}{A character string that will be appended to the front of the output filename (will end with "_sGD.csv").}

\item{xy_file}{A comma delimited file containing columns for individual IDs, X coordinates, and Y coordinates.}

\item{dist_mat}{An NxN (N= sample size) matrix of pairwise landscape distances (Euclidean or effective).}

\item{radius}{The radius of the genetic neighborhood in the same units as \code{dist_mat}.}

\item{min_N}{The minimum sample size per neighborhood for indices to be calculated. NA is returned for neighborhoods < \code{min_N}.}

\item{NS_ans}{Boolean (T or F) answer to whether you want to calculate Wright's neighborhood size.}

\item{GD_ans}{Boolean (T or F) answer to whether you want to calculate genetic diversity indices. This calculation can take a long time depending on how many individuals are in your sample and the \code{radius} of the neighborhood..}

\item{NeEstimator_dir}{Path to the NeEstimator 2.0 directory. NeEstimator 2.0 is required only if NS_ans = T. It can be downloaded from \url{http://molecularfisherieslaboratory.com.au/neestimator-software}.}
}
\value{
A comma delimited text file containing estimates of genetic diversity and neighborhood size for neighborhoods surrounding each sample location.
}
\description{
Calculate spatially explicit indicies of genetic diversity and Wright's neighborhood size (NS).
}
\examples{
#Make sure your paths are correct for your operating system (e.g. in linux, it might be "/home/yourname/Temp")
library(sGD)
setwd("C:/Temp") # the output file will be written to the working directory
genepop_file = system.file("extdata","sGD_demo_genepop.gen",package="sGD")
output_name = "demo"
xy_file = system.file("extdata","sGD_demo_xy.csv",package="sGD")
dist_mat = system.file("extdata","sGD_demo_cdmat.csv",package="sGD")
radius = 16238 # for this demo, units are in meters
min_N = 20
NS_ans = TRUE
GD_ans = TRUE
NeEstimator_dir = "C:/NeEstimator"

sGD(genepop_file,output_name,xy_file,dist_mat,radius,min_N,NS_ans,GD_ans,NeEstimator_dir)

# specify landscape and sGD output
landscape = raster(system.file("extdata","sGD_demo_IBR_landscape.asc",package="sGD"))
sGD_output = read.csv(system.file("extdata","sGD_demo_output_sGD.csv",package="sGD"))
# Convert raster to dataframe for ggplot
landscape.p <- rasterToPoints(landscape)
landscape.df <- data.frame(landscape.p)
colnames(landscape.df) <- c("X", "Y", "Resistance")

# Plot sGD output (Ap is shown here, but explore all sGD outputs) atop the resistance model
library(ggplot2)

ggplot()  +
 geom_raster(data=landscape.df,aes(x=X,y=Y,fill=Resistance),alpha=I(0.5)) +
 scale_fill_gradient(low="black", high="lightgrey") +
 geom_point(data=sGD_output, aes(x=X, y=Y,color=Ap),size=5) +
 scale_color_gradient(low="red", high="green",na.value = "white") +
 theme(panel.grid.major = element_blank(),
       panel.grid.minor = element_blank(),
       panel.background = element_blank())
}

